Rename _pipeline to pipeline
